requirements.txt
telebot
requests
python-dotenv
aiohttp